cambios11abril

+++ TP1/rojas.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include "rojas.h"
 
-float sumaf (float a, float b)
+int sumaf (int a, int b)
 {
     return a + b;
 }
@@ -11,7 +11,7 @@ float sumaf (float a, float b)
 
 
 
-float restaf (float a, float b)
+int restaf (int a, int b)
 {
     return a - b;
 }
@@ -19,15 +19,15 @@ float restaf (float a, float b)
 
 
 
-float divisionf (float a, float b)
+float divisionf (int a, int b)
 {
-    return a / b;
+    return (float) a / b;
 }
 
 
 
 
-float multiplicacionf (float a, float b)
+int multiplicacionf (int a, int b)
 {
     return a * b;
 }
@@ -35,7 +35,7 @@ float multiplicacionf (float a, float b)
 
 
 
-float factorialfa (float a){
+int factorialfa (int a){
 int factorial = 1;
 for (int i=1; i<=a; i++){
     factorial=factorial * i;
@@ -47,7 +47,7 @@ return factorial;
 
 
 
-float factorialfb (float b){
+int factorialfb (int b){
 int factorial = 1;
 for (int i=1; i<=b; i++){
     factorial=factorial * i;
@@ -58,20 +58,20 @@ return factorial;
 
 
 
-int menu (float a, float b)
+int menu (int a, int b)
 {
 int opcion;
     printf (" -----------------------------------------\n");
     printf (" ||||||***** MENU DE OPCIONES *****|||||| \n");
     printf (" -----------------------------------------\n\n");
-    printf ("1) Ingrese primer operando: (A = %.2f)\n",a);
-    printf("2) Ingrese segundo operando: (B = %.2f)\n",b);
+    printf ("1) Ingrese primer operando: (A = %d)\n",a);
+    printf("2) Ingrese segundo operando: (B = %d)\n",b);
     printf("3) Calcular todas las operaciones\n");
-    printf("    a) Calcular la suma %.2f + %.2f\n",a,b);
-    printf("    b) Calcular la resta %.2f - %.2f\n",a,b);
-    printf("    c) Calcular la division %.2f / %.2f\n",a,b);
-    printf("    d) Calcular la multiplicacion %.2f * %.2f\n",a,b);
-    printf("    e) Calcular el factorial de %.2f y de %.2f\n",a,b);
+    printf("    a) Calcular la suma %d + %d\n",a,b);
+    printf("    b) Calcular la resta %d - %d\n",a,b);
+    printf("    c) Calcular la division %d / %d\n",a,b);
+    printf("    d) Calcular la multiplicacion %d * %d\n",a,b);
+    printf("    e) Calcular el factorial de %d y de %d\n",a,b);
     printf("4) Informar resultados\n");
     printf("5) Salir\n\n");
     printf ("Ingrese opcion: \n");
